create ollama chat client
