+++ lib/shared/src/ollama/completions-client.ts
